Added ltoa and ltoa_base. Modified the itoa functions to call their ltoa counterparts

+++ src/ft_digitcount.c
@@ -6,13 +6,13 @@
 /*   By: rhallste <rhallste@student.42.us.org>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/03 13:56:51 by rhallste          #+#    #+#             */
-/*   Updated: 2017/11/03 14:12:59 by rhallste         ###   ########.fr       */
+/*   Updated: 2017/12/05 20:16:36 by rhallste         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/libft.h"
 
-int	ft_digitcount(unsigned int n)
+int	ft_digitcount(unsigned long long n)
 {
 	return (ft_digitcount_base(n, 10));
 }
